# Don't empty dbcache on prune flushes: >30% faster IBD