--- a/src/coins.h
+++ b/src/coins.h
@@ -130,6 +130,9 @@ struct CCoinsCacheEntry
     CCoinsCacheEntry() : flags(0) {}
     explicit CCoinsCacheEntry(Coin&& coin_) : coin(std::move(coin_)), flags(0) {}
     CCoinsCacheEntry(Coin&& coin_, unsigned char flag) : coin(std::move(coin_)), flags(flag) {}
+
+    inline bool IsDirty() const noexcept { return flags & DIRTY; }
+    inline bool IsFresh() const noexcept { return flags & FRESH; }
 };
 
 /**