… src/test/coins_tests.cpp
… src/test/coinscachepair_tests.cpp
… src/test/fuzz/coins_view.cpp
… src/test/fuzz/coinscache_sim.cpp
… src/txdb.h
… src/validation.cpp