@jamesob 2023-08-24
Concept ACK! I'll try to repro the bench results in the next week or so.