@0xB10C 2023-08-30
@andrewtoth Not sure if you have seen https://github.com/bitcoin/bitcoin/pull/20827 which is a different approach but has a similar effect.