@andrewtoth 2023-08-30
@0xB10C indeed I have seen that one. I believe they compliment each other. #20827 reduces IBD time for larger `prune` values and this PR reduces it even more for smaller `prune` values. I should benchmark this on top of that one.
Also, I'm curious how you are performing your benchmarks and generating all those nice graphs. Do you mind sharing what tools you use?