@0xB10C 2023-08-31
> Also, I'm curious how you are performing your benchmarks and generating all those nice graphs. Do you mind sharing what tools you use?

At the time I had a spare and otherwise idle machine set up I could use. I don't have the bash script at hand but IIRC I had a merge base (MB)  and a pull-request (PR) binary I'd run with different `prune` and `dbcache` configurations. After each run, the script would rename and move the debug log (with extra `-debug=coindb -debug=prune`) to a separate folder. After all runs were done, I loaded the debug logs into the following Jupyter notebooks to parse the debug logs and to generate the graphs. For anyone revisiting this, these are the relevant comments https://github.com/bitcoin/bitcoin/pull/20827#issuecomment-1009016428, https://github.com/bitcoin/bitcoin/pull/20827#issuecomment-1017603590, and https://github.com/bitcoin/bitcoin/pull/20827#issuecomment-1023108273 for my setup and results.

I've uploaded the jupyter nodebooks here https://gist.github.com/0xB10C/89c2903290cfb1840792d41dcd079646. The first was used for a partial IBD from 500k-600k and the second one with a full IBD. Both times syncing from a local node on the same host. The notebooks expect the debug log files with the name `debug_{binary}_{cset}_{run}.log` where {binary} is either MB or PR, {cset} is an integer specifying the configuration set (pruning and dbcache parameter), and {run} is an integer numbering the run of this binary-cset combination if I did multiple. Feel free to remix and share in any way you want. Keep in mind that parsing logs is quite fragile as log message format can change. Feel free to reach out if there are questions.