@jamesob 2023-09-20
Oddly, I benched this on a 30_000 block contemporary region with an `-dbcache=4000 -prune=8800` and saw no difference...

I went to investigate the debug.log to ensure that we actually saw some prune events, but realized my `-debug` settings aren't sufficient to log those. Kind of surprising that we wouldn't log prune events by default.

I'll re-run with different log settings.

@andrewtoth is there any reason you can think of that the parameters of this bench wouldn't capture the expected performance improvements?

### commands index
|                      bench name                       |                                                                                                           command                                                                                                           |
|-------------------------------------------------------|-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
| ibd.local.range.dbcache=4000-prune=8800.667200.697200 | `bitcoind -dbcache=4000 -prune=8800 -debug=coindb -debug=bench -listen=0 -connect=0 -addnode=127.0.0.1:8888 -prune=9999999 -printtoconsole=0 -assumevalid=000000000000000000176c192f42ad13ab159fdb20198b87e7ba3c001e47b876` |


### #20827 vs. $mergebase (absolute)
|                              bench name                               |  x  |           #20827           |         $mergebase         |
|-----------------------------------------------------------------------|----:|----------------------------|----------------------------|
| ibd.local.range.dbcache=4000-prune=8800.667200.697200.total_secs      |   3 | 4527.4863 (± 11.8146)      | 4550.3649 (± 30.9606)      |
| ibd.local.range.dbcache=4000-prune=8800.667200.697200.peak_rss_KiB    |   3 | 4804738.6667 (± 1283.9993) | 4804654.6667 (± 1206.0573) |
| ibd.local.range.dbcache=4000-prune=8800.667200.697200.cpu_kernel_secs |   3 | 211.0633 (± 0.8949)        | 211.6333 (± 0.3866)        |
| ibd.local.range.dbcache=4000-prune=8800.667200.697200.cpu_user_secs   |   3 | 26388.7033 (± 6.4040)      | 26439.3633 (± 36.6065)     |


### #20827 vs. $mergebase (relative)
|                              bench name                               |  x  | #20827 | $mergebase |
|-----------------------------------------------------------------------|----:|-------:|-----------:|
| ibd.local.range.dbcache=4000-prune=8800.667200.697200.total_secs      |   3 |   1.00 |      1.005 |
| ibd.local.range.dbcache=4000-prune=8800.667200.697200.peak_rss_KiB    |   3 |   1.00 |      1.000 |
| ibd.local.range.dbcache=4000-prune=8800.667200.697200.cpu_kernel_secs |   3 |   1.00 |      1.003 |
| ibd.local.range.dbcache=4000-prune=8800.667200.697200.cpu_user_secs   |   3 |   1.00 |      1.002 |