@jamesob 2023-09-20
Bench running now, should have some results in the next day or so.
```shell
$ ( bitcoinperf --no-teardown bench-pr --num-blocks 30_000 --run-id no-flush-on-prune \
    --bitcoind-args='-dbcache=4000 -prune=550' --run-count 3 28280 && \
    pushover 'Bench for andrewtoth SUCCEEDED!' ) || \
  pushover 'Bench for andrewtoth failed'
```