@jamesob 2023-09-27
Local IBD of 30_000 blocks with `-prune=550` from a Taproot-enabled height didn't show any difference. I didn't have `-log=prune` enabled, so my bench debug.log was basically useless. Rerunning...