@martinus 2023-12-05
I didn't think this through yet, but this might be an alternative way to achieve the same thing: Instead of the double linked list, split the `cacheCoins` map up into 4 different maps, one map for each flag setting, e.g. as a `std::array<CCoinMap, 4> m_cache_coins_per_flag;`. Index is the flag, e.g. `m_cache_coins_per_flag[FRESH | DIRTY]` to get the map with all entries that are fresh & dirty. Instead of setting a flag & linking the list, use the unordered_map's `auto  node = extract()` and `insert(std::move(node))` to move the entry to the corresponding map. This only relinks the unordered_map node and causes no allocation, pointers to the entry stay valid. 

This approach would have a few advantages:
* Less memory usage: no `m_flag´ member is needed that actually required 8 bytes due to alignment, no 2 pointers needed for interlinking, and multiple maps means overall bucket sizes will stay smaller.
* processing the maps might be faster: iteration is fast because e.g. skipping non-dirty entries is trivial (they are in a different map). There might be a hashing overhead though from `extract()` and `insert()`.

For this to work, I think all 4 maps need to have the same `m_cache_coins_memory_resource`. Moving nodes that use the memory resource between maps is untested though; but I'm pretty sure it should work out of the box.