@andrewtoth 2023-12-05
@martinus something similar was mentioned in https://github.com/bitcoin/bitcoin/pull/15265#issuecomment-457720636. However, with your approach every access would require 4 lookups.