@paplorinc 2024-07-09
The [benchmarks](https://corecheck.dev/bitcoin/bitcoin/pulls/28280) indicate that `BlockToJsonVerboseWrite` was slowed down

<img src="https://github.com/bitcoin/bitcoin/assets/1841944/8b5d5299-a4c2-4b26-9a74-44dd2504a20c">

I ran a before/after against `master`, found only a minor (4%) diff - is this expected?
> make -j10 && ./src/bench/bench_bitcoin --filter=BlockToJsonVerboseWrite --min-time=10000


before:
|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|       27,574,397.06 |               36.27 |    0.2% |     11.03 | `BlockToJsonVerboseWrite`
|       27,586,627.40 |               36.25 |    0.1% |     11.01 | `BlockToJsonVerboseWrite`
|       27,638,839.92 |               36.18 |    0.2% |     10.99 | `BlockToJsonVerboseWrite`

after:
|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|       28,573,375.00 |               35.00 |    0.2% |     11.06 | `BlockToJsonVerboseWrite`
|       28,531,835.59 |               35.05 |    0.1% |     11.08 | `BlockToJsonVerboseWrite`
|       28,602,410.56 |               34.96 |    0.1% |     11.09 | `BlockToJsonVerboseWrite`