@vostrnad 2024-08-05
Benchmark ACK (that's not a thing but it totally should be)

I benchmarked the latest revision (28f0dd2d60d0f9a8d8a5aa0ed4a5eb1bd4f605bf) against the merge base (9774a958b501a6d439a734e18b29e04f59f973f6) with a few runs to block 400,000, syncing from a local node, with pruning enabled and all other settings at default. With `dbcache=450` it finished 9.4% faster, `dbcache=4096` was 14.3% faster.

**dbcache=450:**

master: 3203s
pr: 2901s (9.4% speedup)

![dbcache450](https://github.com/user-attachments/assets/b3b17b50-f6d1-49ba-96b8-7cc71453fcc6)

**dbcache=4096:**

master: 3241s
pr: 2778s (14.3% speedup)

![dbcache4096](https://github.com/user-attachments/assets/af347efb-12ba-4bca-a070-d08ee7edb024)